+++ terminaltexteffects/effects/effect_expand.py
@@ -9,17 +9,18 @@ Classes:
 from __future__ import annotations
 
 from dataclasses import dataclass
+from typing import cast
 
-from terminaltexteffects import Color, EffectCharacter, EventHandler, Gradient, easing
+from terminaltexteffects import Color, EffectCharacter, EventHandler, Gradient, Scene, easing
 from terminaltexteffects.engine.base_config import (
     BaseConfig,
     FinalGradientDirectionArg,
-    FinalGradientFramesArg,
     FinalGradientStepsArg,
     FinalGradientStopsArg,
 )
 from terminaltexteffects.engine.base_effect import BaseEffect, BaseEffectIterator
 from terminaltexteffects.utils import argutils
+from terminaltexteffects.utils.graphics import ColorPair
 
 
 def get_effect_resources() -> tuple[str, type[BaseEffect], type[BaseConfig]]:
@@ -43,8 +44,6 @@ class ExpandConfig(BaseConfig):
             provided, the characters will be displayed in that color.
         final_gradient_steps (tuple[int, ...] | int): Tuple of the number of gradient steps to use. More steps will
             create a smoother and longer gradient animation. Valid values are n > 0.
-        final_gradient_frames (int): Number of frames to display each gradient step. Increase to slow down the
-            gradient animation.
         final_gradient_direction (Gradient.Direction): Direction of the final gradient.
 
     """
@@ -56,8 +55,7 @@ class ExpandConfig(BaseConfig):
         epilog=(
             f"{argutils.EASING_EPILOG}"
             "Example: terminaltexteffects expand --final-gradient-stops 8A008A 00D1FF FFFFFF --final-gradient-steps 12 "
-            "--final-gradient-frames 5 --final-gradient-direction vertical --movement-speed 0.35 "
-            "--expand-easing IN_OUT_QUART"
+            "--final-gradient-direction vertical --movement-speed 0.35 --expand-easing IN_OUT_QUART"
         ),
     )
     expand_easing: easing.EasingFunction = argutils.ArgSpec(
@@ -93,11 +91,6 @@ class ExpandConfig(BaseConfig):
         "create a smoother and longer gradient animation."
     )
 
-    final_gradient_frames: int = FinalGradientFramesArg(
-        default=5,
-    )  # pyright: ignore[reportAssignmentType]
-    "int : Number of frames to display each gradient step. Increase to slow down the gradient animation."
-
     final_gradient_direction: Gradient.Direction = FinalGradientDirectionArg(
         default=Gradient.Direction.VERTICAL,
     )  # pyright: ignore[reportAssignmentType]
@@ -119,7 +112,7 @@ class ExpandIterator(BaseEffectIterator[ExpandConfig]):
         """
         super().__init__(effect)
         self.pending_chars: list[EffectCharacter] = []
-        self.character_final_color_map: dict[EffectCharacter, Color] = {}
+        self.character_final_color_map: dict[EffectCharacter, ColorPair] = {}
         self.build()
 
     def build(self) -> None:
@@ -133,7 +126,15 @@ class ExpandIterator(BaseEffectIterator[ExpandConfig]):
             self.config.final_gradient_direction,
         )
         for character in self.terminal.get_characters():
-            self.character_final_color_map[character] = final_gradient_mapping[character.input_coord]
+            if self.terminal.config.existing_color_handling == "dynamic":
+                self.character_final_color_map[character] = ColorPair(
+                    fg=character.animation.input_fg_color,
+                    bg=character.animation.input_bg_color,
+                )
+            else:
+                self.character_final_color_map[character] = ColorPair(
+                    fg=final_gradient_mapping[character.input_coord],
+                )
         for character in self.terminal.get_characters():
             character.motion.set_coordinate(self.terminal.canvas.center)
             input_coord_path = character.motion.new_path(
@@ -156,13 +157,42 @@ class ExpandIterator(BaseEffectIterator[ExpandConfig]):
                 0,
             )
             character.motion.activate_path(input_coord_path)
-            gradient_scn = character.animation.new_scene()
-            gradient = Gradient(final_gradient.spectrum[0], self.character_final_color_map[character], steps=10)
-            gradient_scn.apply_gradient_to_symbols(
-                character.input_symbol,
-                self.config.final_gradient_frames,
-                fg_gradient=gradient,
-            )
+            gradient_scn = character.animation.new_scene(sync=Scene.SyncMetric.DISTANCE)
+            if self.terminal.config.existing_color_handling == "dynamic":
+                fg_gradient = (
+                    Gradient(final_gradient.spectrum[0], character.animation.input_fg_color, steps=10)
+                    if character.animation.input_fg_color
+                    else None
+                )
+                bg_gradient = (
+                    Gradient(final_gradient.spectrum[0], character.animation.input_bg_color, steps=10)
+                    if character.animation.input_bg_color
+                    else None
+                )
+                if fg_gradient or bg_gradient:
+                    gradient_scn.apply_gradient_to_symbols(
+                        character.input_symbol,
+                        1,
+                        fg_gradient=fg_gradient,
+                        bg_gradient=bg_gradient,
+                    )
+                else:
+                    gradient_scn.add_frame(
+                        character.input_symbol,
+                        1,
+                        colors=ColorPair(),
+                    )
+            else:
+                gradient = Gradient(
+                    final_gradient.spectrum[0],
+                    cast("Color", self.character_final_color_map[character].fg_color),
+                    steps=10,
+                )
+                gradient_scn.apply_gradient_to_symbols(
+                    character.input_symbol,
+                    5,
+                    fg_gradient=gradient,
+                )
             character.animation.activate_scene(gradient_scn)
 
     def __next__(self) -> str: